Added Customer Details page

+++ API/controllers/customerBillControllers.js
@@ -6,11 +6,20 @@ import Customer from "../models/customerModel.js";
 export const addCustomerBill = async (req, res) => {
   try {
     // request body se customerType, customerName aur items ko destructure kar rahe hain
-    const { customerType, customerName, items } = req.body;
+    const { customerType, customerName, items, date } = req.body;
 
     // agar ye tino fields body me na milein to error return kar do
     if (!customerType || !customerName || !items) {
       return errorHandler(res, 400, "Missing Fields");
+    }
+
+    let finalDate;
+    if (date) {
+      // frontend se string aayi hai to wahi store karo
+      finalDate = date;
+    } else {
+      // agar nahi aayi to backend khud string banaye
+      finalDate = new Date().toISOString().split("T")[0];
     }
 
     // updated items ki array banayi jisme final processed items push karenge
@@ -21,7 +30,7 @@ export const addCustomerBill = async (req, res) => {
     // har item ke liye loop chalaya
     for (let i of items) {
       // product ko database se find kar rahe hain id ke basis par
-      const product = await Product.findById(i.product);
+      const product = await Product.findOne({ productName: i.product });
       if (!product) {
         // agar product na mile to error return
         return errorHandler(res, 400, "Product not found");
@@ -56,7 +65,7 @@ export const addCustomerBill = async (req, res) => {
 
       // updatedItems array me processed item push karna
       updatedItems.push({
-        product: i.product,
+        product: product._id,
         quantity: i.quantity,
         price,
         total: itemTotal,
@@ -77,10 +86,10 @@ export const addCustomerBill = async (req, res) => {
     }
 
     let customerId = null;
-    if(customerType === "regular"){
-      const customer = await Customer.findOne({ customerName })
-      if(customer){
-        customerId = customer._id
+    if (customerType === "regular") {
+      const customer = await Customer.findOne({ customerName });
+      if (customer) {
+        customerId = customer._id;
       }
     }
 
@@ -92,6 +101,7 @@ export const addCustomerBill = async (req, res) => {
       items: updatedItems,
       billNumber: newBillNumber,
       totalAmount,
+      date: finalDate,
     });
 
     // bill ko database me save kia
@@ -104,7 +114,7 @@ export const addCustomerBill = async (req, res) => {
       if (customer) {
         // uske bills array me savedBill ka id push karna
         customer.bills.push(savedBill._id);
-        customer.balance += totalAmount
+        customer.currentBalance += totalAmount;
         await customer.save();
       }
     }
@@ -114,7 +124,6 @@ export const addCustomerBill = async (req, res) => {
     return errorHandler(res, 400, error?.message);
   }
 };
-
 
 // get all customers bills
 export const getAllCustomerBills = async (req, res) => {
@@ -148,54 +157,73 @@ export const getAllCustomerBills = async (req, res) => {
   }
 };
 
-
 //get singleCustomerBill
 export const getSingleCustomerBill = async (req, res) => {
   try {
-    const { id } = req.params
+    const { id } = req.params;
 
-    const singleBill = await CustomerBill.findById(id).populate("items.product")
-    if(!singleBill){
-      return errorHandler(res, 400, "Bill not found by given ID")
+    const singleBill = await CustomerBill.findById(id).populate(
+      "items.product"
+    );
+    if (!singleBill) {
+      return errorHandler(res, 400, "Bill not found by given ID");
     }
 
-    return successHandler(res, 200, "Bill Found", singleBill)
+    return successHandler(res, 200, "Bill Found", singleBill);
   } catch (error) {
-    return errorHandler(res, 400, error?.message)
+    return errorHandler(res, 400, error?.message);
   }
-}
-
+};
 
 export const getSingleCustomerBills = async (req, res) => {
   try {
     const { id } = req.params;
 
-    const bills = await CustomerBill.find({ customerId : id })
-    if(!bills || bills.length === 0){
-      return errorHandler(res, 400, "Customer Bills Not found")
+    const bills = await CustomerBill.find({ customerId: id });
+    if (!bills || bills.length === 0) {
+      return errorHandler(res, 400, "Customer Bills Not found");
     }
 
-    return successHandler(res, 200, "Single customer Bills fetched", bills)
+    return successHandler(res, 200, "Single customer Bills fetched", bills);
   } catch (error) {
-    return errorHandler(res, 400, error?.message)
+    return errorHandler(res, 400, error?.message);
   }
-}
-
-
+};
 
 //Delete customer bill handler
 export const deleteCustomerBill = async (req, res) => {
   try {
     const { id } = req.params;
 
-    const deletedBill = await CustomerBill.findByIdAndDelete(id)
-
-    if(!deletedBill){
-      return errorHandler(res, 400, "Bill not found!")
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      return errorHandler(res, 400, "Invalid Vendor Bill ID");
     }
 
-    return successHandler(res, 200, "BIll Deleted Successfully", deletedBill)
+    const deletedBill = await CustomerBill.findByIdAndDelete(id);
+
+    if (!deletedBill) {
+      return errorHandler(res, 400, "Bill not found!");
+    }
+
+    for(let item of deletedBill.items){
+      let product = await Product.findById(item.product)
+      if(product){
+        product.quantity += item.quantity
+        await product.save()
+      }
+    }
+
+    let customer = await Customer.findById(deletedBill.customerName)
+    if(customer){
+      customer.customerBills = customer.customerBills.filter((bill) => !bill.equals(deletedBill._id))
+      customer.currentBalance -= deletedBill.totalAmount;
+      customer.totalTurnover -= deletedBill.totalAmount
+      await customer.save()
+    }
+
+
+    return successHandler(res, 200, "BIll Deleted Successfully", deletedBill);
   } catch (error) {
-    return errorHandler(res, 400, error?.message)
+    return errorHandler(res, 400, error?.message);
   }
-}
+};
